[bioshell-pdb] SecondaryStructure class reimplemented
Now it contains a Vec<SecondaryStructureTypes>. Secondary structure codes (integers from 1 to 12) are no longer used.

+++ bioshell-pdb/tests/test_pdb_atom.rs
@@ -19,7 +19,7 @@ mod tests {
         assert_eq!(atom.occupancy, 1.0);
         assert_eq!(atom.temp_factor, 0.0);
         assert_eq!(atom.element, Some(String::from("C")));
-        assert_eq!(atom.secondary_struct_type, SecondaryStructureTypes::Coil as u8);
+        assert_eq!(atom.secondary_struct_type, SecondaryStructureTypes::Coil);
         assert_eq!(atom.charge, None);
     }
